Rewrite ubignum_add() and add macro max, min
Put the computation result in variable ans, which is allocated when
there is pointer aliasing. In this way, we no longer have to restore @a
and @b in the case that resize() failed.

Current issue is that the bit shift of a unsigned __int128 number does
not result in what we expect.