Fix resize(), add sub(), and remove duplicate()
ubignum_resize():
We should reallocate (*N)->data instead of *N.
Our static analysis tool treats realloc() with size = 0 as a memleak,
so I have no choice to write the free() functions.

ubignum_sub():
It can only result in unsigned number.
The addition cannot apply ubignum_add() because we want to discard the
highest carry and we do not need the size checkings.

ubignum_duplicate():
Since it conflicts with ubignum_init() and its usage confuses, I decided
to remove this function.