Free old output when ptr aliasing
Pointer aliasing is a challenging issue.

This patch tries to solve problem for like add(a, b, &a). The old @a
will be correctly freed in add().

However when there are different pointers a and b pointing to the same
location, and call by add(a, b, &a), then @b would point to somewhere
freed. This has not been fixed.