Add zero(), divby_ten() and fix bug in lef_shift()
ubignum_zero():
Help to reset (*out) in each ubn operations.
It has not added in all operations.

ubignum_divby_ten():
Implemented by bit shift and subtraction.
When we want to print bignum result in decimal such that human can
understand, we need this function to help conversion.
There is still bug in this function to be solved. See _debug.c.

ubignum_left_shift():
Fix an essential bug!
In the previous patch, I thought that left-shifted by ubn_unit_bit is
equivalent to shifted by 0. But that is wrong. What it should do is set
the number to 0.