Redesign ans resize in add() and modify compare()
ubignum_add():
To prevent calling resize() during addition, we compute a maximum needed
size and up the capacity to it.

ubignum_compare():
Now use size to determine order of the numbers, so the highest chunk of
any ubn should not be 0.

ubignum_2decimal():
The computation of digit is improved to fit the num.