Rewrite divby_ten() to a clearer version
ubignum_divby_ten():
Allocate everything needed at the start of the function. Thus, in case
there is allocation failure, we would not waste time at computation.
Since we have to allocate space for the answer, we do not have to care
about there is aliasing or not.

ubignum_init():
Use calloc() to prevent calling zero().