Rewrite and fix mult(), mult_add() and add iszero()
mult():
Pointer aliasing no longer causes a special case.
My previous thought about carry when  computing the partial product was
wrong and complicated. Fix and rewrite to a simpler and correct way.
The concept is descibed in comments.

mult_add():
In order to do it faster, remove most of the checkings and resize.
Let mult() do the neccessary checking and resize.

iszero():
Since the zero in ubn is a little bit special, add a function for
checking 0 easier.

TODO:
1. The reallocation policy in ubignum_add()
2. ubignum_divby_ten() with quo = 0