Add copy(), fix mult(), mult_add() and left_shift()
Add ubignum_copy().

ubignum_mult():
Forgot to free pprod. That caused memory leak.

ubignum_mult_add():
Since the size is guaranteed in mult() and no resize is perfomed, we
should not introduce the +1 when using (*out)->size.

ubignum_left_shift():
The condition for a = 0 is now considered.

Besides, the array size for computing fast doubling is now reduced to a
constant number 5.