Divide bignum.h to base.h, ubignum.c and ubignum.h
Architecture and compiler dependencies are collected in base.h. In this
way, the ubignum implementation can be in a more abstract level.

Most check to the ubn_t pointers at the beginning of the operation
functions are removed. Because these checks cannot protect the program
against the uninitialized input but only protect it from very few
situations.

The prototype of ubignum_init() is changed, to a more intuitive form.
And a capacity argument is required to init a bignum with appropriate
space without a combind calling to ubignum_recap().

The prototype of ubn_unit_add() is changed as well.

resize() is renamed to recap() to prevent confusion.