Modify Makefile and fibdrv.c
Makefile:
Since we have another source file, we modify Makefile to link all the C
source files and rename TARGET_MODULE to fibdrv_main.

fibdrv.c:
A severe memory leakage in fib_read() is fixed.
fib_sequence() is modified to provide information with printk(), and
the redundant @index is removed.