Introduce a memory pool to divby_ten()
When we intend to convert our ubignum to 10-base, ubignum_2decimal()
calls ubignum_divby_ten() many many times. At every time, divby_ten()
allocated memory to finish its job. This is not efficient.
Thus, we introduce a struct which containing the variables needed for
divby_ten(), and it would be allocated only once in ubignum_2decimal().
This strategy is expected to greatly improve the timing consumption.