Fix ubignum_left_shift() when pointer aliasing
In b3f6b0, we modified left_shift() to do copy from lower to higher.
However, this brought the wrong result if there is pointer aliasing.
We should not modify the data that we are going to use. Therefore, if
we want to do the left-shift operation in-place, we have to start from
the higher part.