Remove constant-qualifiers at some bignum ops
In case there is pointer aliasing, the pointers to the const-qualified
object (ubn_t) cause undefined behaviour. According to C99 standard,
the attempt that tries to modify const-qualified object by their lvalue
with non-const-qualified type is undefined.
Though there was no undesired result generated by the program, we have
to fix this issue to prevent the possibly wrong on other platforms.
We leave the const qualifiers for those functions that do not modify
memory contents.