Replace divby_ten() with divby_superten()
We implemented ubignum_2decimal() by dividing 10 again and again.
In order to speed up this conversion, we may do the division with a
power of ten, then we can get several digits in each iteration.

The logic of divby_superten() function is basically identical to the
divby_ten().

For the remainders, we use snprintf() to obtain strings.
The strings have to be reverted since we need to do the overall reverse
at last. And it is because the first digit-block from the division
iteration is the lowest significant.